Revert switching the openSUSE desktoptheme to dark colors
Switching the default to a dark theme is a rather invasive change, as it hits
existing installations. On top of the change itself, it appears like some
Plasma caches aren't invalidated and after installation there's an unnusable
mixup of old and new colors being used, leading to e.g. unreadable text.